Fix saving versions when safe mode is enabled for Monogo.
The save being triggered in the before_save callback was leading 2 save
attempts for each record - once during the before_save and then once
again during the normal save process. This second, normal save was then
failing because a record already existed in Mongo. When safe mode is not
enabled, Mongo's second save was silently failing, but with safe mode
turned on, this error means each save action will result in a failure.